requirements.txt
asgiref
backports.entry-points-selectable
black
braintree
cachetools
certifi
charset-normalizer
click
distlib
Django
django-cors-headers
django-environ
django-filter
djangorestframework
filelock
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
mypy-extensions
numpy
oauthlib
pandas
pathspec
paytmchecksum
Pillow
pipenv
platformdirs
protobuf
psycopg2-binary
pyasn1
pyasn1-modules
pycrypto
pycryptodome
pyparsing
python-dateutil
pytz
razorpay
regex
requests
requests-oauthlib
rsa
six
sqlparse
tomli
typing-extensions
uritemplate
urllib3
virtualenv
virtualenv-clone